simplified descriptors
