+++ src/main/java/backend/model/serviceprovider/GenericServiceProvider.java
@@ -16,6 +16,7 @@ import org.springframework.core.type.filter.AssignableTypeFilter;
 import org.springframework.core.type.filter.TypeFilter;
 
 import ro.fortsoft.pf4j.ExtensionPoint;
+import backend.model.Descriptor;
 import backend.model.GlobalPersistenceUnit;
 import backend.model.job.JobExecutor;
 import backend.model.result.Result;
@@ -27,62 +28,11 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public abstract class GenericServiceProvider implements ExtensionPoint, ServiceProvider{
 	
-	public static class ServiceProviderDescriptor
+	public static class ServiceProviderDescriptor extends Descriptor<GenericServiceProvider>
 	{
-		private Class<GenericServiceProvider> m_classDescriptor;
-		
-		@JsonProperty("commonName")
-		private String m_commonName;
-		
-		@JsonProperty("identifier")
-		private String m_identifier;
-		
-		public ServiceProviderDescriptor(Class<GenericServiceProvider> clazz)
+		public ServiceProviderDescriptor(Class<GenericServiceProvider> providerClass)
 		{
-			m_classDescriptor = clazz;
-			try
-			{
-				String identifier = m_classDescriptor.getCanonicalName();
-				MessageDigest messageDigest;
-				messageDigest = MessageDigest.getInstance("SHA");
-				messageDigest.update(identifier.getBytes());
-				identifier = String.format("%040x", new BigInteger(1, messageDigest.digest()));
-		    	
-				identifier(identifier);
-			}
-			catch(NoSuchAlgorithmException e)
-			{
-				e.printStackTrace();
-			}
-		}
-		
-		public Class<GenericServiceProvider> classDescriptor()
-		{
-			return m_classDescriptor;
-		}	
-		
-		@JsonProperty("commonName")
-		public void commonName(String name)
-		{
-			m_commonName = name;
-		}
-		
-		@JsonProperty("commonName")
-		public String commonName()
-		{
-			return m_commonName;
-		}
-		
-		@JsonProperty("identifier")
-		public void identifier(String name)
-		{
-			m_identifier = name;
-		}
-		
-		@JsonProperty("identifier")
-		public String identifier()
-		{
-			return m_identifier;
+			super(providerClass);
 		}
 	}
